fix unique key warning for ExternalLink.render
For lists of elements, react needs a unique `key` to determine what
gets re-rendered even when order changes. However, the `key` is
read-only and can be set only when creating the element. The way we
manipulated the children list here thus didn't work without a warning.
We now use the `props.children` as is via JSX and create the optional
icon in JSX as well. Since there is now no list of elements left that
we change, react can take care of the keys itself.