Switching the layout styles from .wp-primary to .wp-content
Not all the elements on the screen live inside .wp-primary. Things like
notices are children to the primary element. This causes some hassles
with the new sidebar layout. To make it easier, lets apply our layout
styles to .wp-content. Along the way I moved the sidebar width to a
variable.