Framework: create a new layout state for sans-sidebar layout.
Use the new noSidebar layout state on the welcome screen.

Making the new noSidebar state actually do something — in this case, it
removes the extra padding thats not needed without the sidebar.

Switching everything to use to new noSidebar layout.

Making sure the padding on wp-content is 8px on mobile.