Show a message in browser while still compiling
Currently when we try to open http://calypso.localhost:3000/ while
webpack is still compiling we see a blank page and a message in the
console. Unfortunately the message in the console isn’t 100% clear and
we can’t expect all people to look there.

With this change if somebody opens http://calypso.localhost:3000/ they
will see a welcome message explaining what’s going on and what to do.
The page will refresh itself every 5 seconds, so the old behavior of
the page loading automatically is kinda preserved.