Show login interstitial in development mode
If a user is logged out they can’t do much, even with the dev version.

This shows them a screen inviting them to login or register.

Since the page is hosted in `/devdocs` we enable it for logged out users.
It didn’t make sense to create a brand new section just for this.