fix unique key warning for BulkSelect example
Add a key based on the element's position in the internal (state)
array so react can distinguish the different created `label` elements.

As per the React docs (reconciliation.html#keys), keys should be
"stable, predictable, and unique." Importantly, keys need only be
"unique among its siblings, not globally unique." In this case, the
data source in this example is a static array of elements that is
never changed or reordered. Additionally, JavaScript's `map` goes
through the array elements in order. Using the position in the list
therefore makes for a suitable key.

Note that this should be considered case-by-case. In non-example code,
elements will likely not come from a static, hard-coded array.