Build: Install pre-commit githook on `make run`
The rationale for this is that we'd like more exposure for the GPLv2
contribution notice. Installing them on `make run` seems like a good way
to do this.

The way that `bin/pre-commit` is set up now, this
will also force the running of i18nlint and eslint checks. Which is a
Good Thing in any case.

Of course, these checks can be bypassed using `git commit --no-verify`