fix "failed propType" warning for BulkSelect
In the BulkSelect example and component, we use checkboxes but handle
interaction through our own handlers (bound with onClick) instead of
using React's onChange mechanism. React complained that those
checkboxes would then need to be marked `readOnly`. This isn't super
intuitive, but we are actually using the checkboxes just for their
presentations, not for their built-in interactions or for recording a
value -- all of that happens in our code. In the example, the
checkboxes are visualizations for the state we hold in the `elements`
array. In the component, the checkbox is a visualization of its props
(all selected, some selected, none selected).