Reader: Pull in stores that accept actions when pre-loading.
The Reader preloads all of a user's subscriptions when you hit any Reader URL. This works pretty well, but we have to ensure that all of the stores that want to listen for and act upon actions fired by the preload are spun up and listening.

This patch teaches the controller how to load all of those stores before kicking off the preload.