Signup: determine user step by `providesToken` instead of by name
When a user is logged-in, the `user` step of any flow is removed because the
user already has an account and does not need to create one. To do this, the
flow logic previously compared the name of the step to `user`. This can be
problematic because there are cases where a different `user` step will be used,
and the string comparison will be invalid.

This changes the logic to look for the `providesToken` property name on the step
config object, which should be present only in `user` steps.