Cast category IDs to strings for compat reasons
Some versions of the API use ctype_digit to validate numbers. ctype_digit returns false for integers between -128 and 255. Some versions of Calypso (namely desktop) send category IDs as integers which means that if the category ID is between 1 and 255, it will always fail to be added to the post.

By forcing it to a string, we can make it always work.