Adds a new action-only poller
This polling component can wrap another component in a polling mechanism
that runs a given function at a given interval and will optionally stop
polling if the component enters a hidden visibility state to save on
bandwidth.

If specified to pause when hidden, there is an optional parameter
describing whether or not the action should fire immediately upon coming
back into view. If not, it will fire after the given interval.

This component was inspired by the data-poller component but attempts to
be a more lightweight and readable component.

Fix poller to work

When I wrote the first draft of this component, I was doing so on a
plane without Internet access and I did not test it out. Not
surprisingly, it didn't work. This revision updates it to work the way
the original code was intended.

This does not change the functionality of the code, merely the syntax,
as the way I thought I might be able to apply the fat-arrow function
syntax didn't work in practice.

Renamed ActionPoller -> Interval

Refactor the interval to aggregate actions

This is a significant change from my earlier proposal. In this
iteration, we use a stateful backing called `runner.js` which both
tracks which actions need to be executed and when. It coalesces
operations so that every action that needs to occur, for example, every
minute, does so together every minute instead of being scattered about.

Currently there are only a few intervals available _by design_ to keep
things simple and encourage aggregation. Further interval groups should
be added only with a proper reason why the existing set is insufficient.

The React component in this iteration is mostly an interface to the
underlying `runner.js`, but handles all of the visibility issues.
Further, the `leading` property has been removed because this interval
makes no guarantees about when exactly that the callbacks will execute,
only that they will do so on the specified interval. Probably the
default behavior ought to be that the callbacks get called initially
after being added to the list.

fixup! Refactor the interval to aggregate actions

fixup! Refactor the interval to aggregate actions

Interval: Ease on the return for Immutable.Map.forEach

When I wrote the code for the `forEach` handler I noticed a warning in
the documentation for Immutable.js that said any iteration returning
`false` would stop iteration. Because of this, I explicitly returned
true to make sure we went through the whole list.

Today I had a chance to test out the behavior and figure out exactly
what happens and so I have rewritten the code to make a better use of
this rule.

 - If the function inside of `forEach` returns === `false` then it will
   stop iteration. `undefined` and `null` do not terminate iteration.

 - Because we still pass of execution to `onTick()` which understandably
   could return `false`, I'm leaving in a guard around that call.

Interval: refactor tests to obviate `Interval.get()`

Previously the tests introspected the interval state object to make sure
that the appropriate behaviors were reflected in the data structure.
Practically speaking, however, the tests should be more focussed on the
result of those behaviors, not on how the code internally organizes
them.

The tests have thus been improved by decoupling from the library code
itself through the use of mutable callbacks that announce whether or not
they have run by modifying an object in the test runners. See
`nudgeObject()`

Further, the fake timers from **Sinon** have been used to help eliminate
these dependencies.

fixup! Interval: refactor tests to obviate `Interval.get()`

Interval: Update documentation

Interval: Pass props to children, update README, test `<Interval/>`

Previously the React component part of this feature was untested. In
this commit it is properly tested. To help with testing, the new
`enzyme` library is pulled in, making the React testing significantly
easier.

The README has been updated.

Interval: Move time period definitions into runner.js

As a solution to defining these values in two places I have moved them
into `runner.js`. Initially I tried to keep them in `index.js`, but
because there would have created a circular dependency I placed them
back in `runner.js`.

Of note, I think that I should have been able to re-export them at the top of
`index.js` without having a separate import/export block, but this left
the code without their definitions, so I had to import first, then
export. Additionally, the re-export syntax here isn't technically the
re-export syntax.

```js
// re-exporting
export { EVERY_SECOND } from './runner';
```

The reason I left out the `from './runner'` is because it's not needed
here and I wasn't sure what kinds of implications it would bring. ES6
officially supports the re-exporting and syntax I was to use, but Babel
actually transpiles this all back to some form of CommonJS module,
making the actual output a bit hazy to me.

For now, however, we have the definitions in a single place,
most-closely tied to core that _should_ define them, and all the tests
pass.

Interval: Fix typo in README.md

Interval: Use optional arg in `setTimeout` instead of creating closure

Previously we were creating a closure to get the argument into
`executePeriodActions` in the `setTimeout` call but this has been
changed because `setTimeout` accepts an optional sequence of arguments
to pass to the callback. This is favorable because it's less
resource-heavy.

Interval: Add clarifying comment about browser visibility

Interval: No need to return unused value in `storeNewAction`

Interval: Add documentation to runner.js

Interval: Return `null` when no children present.

Previously if there were no children the Interval would return
`<span/>` instead of `null`. Now it returns `null`

To accomodate this the test suite had to be updated because of what
appear to be issues with the library handling `null` values. For now,
those instances have been changed to include an empty `<div />` just for
the tests themselves.