Live Chat: Hide the X link that lets people close the chat box.
When you click this, the chat goes away for the remainder of your session and there's not a good way to get it back. We've overridden the standard Olark UI to make the floating box less obtrusive anyway, so the need to X out is rendered unnecessary by the unobtrusive UI we now show when you click the little inverted carat to minimize the chat box.

This change makes it harder for users to accidentally dismiss the chat box for the duration of their session and have difficulty initiating or resuming a chat.