Signup: Verticals Survey: Add verticalSurvey test
The Verticals survey is a signup test wherein a new user is asked to categorize
their soon-to-be new site using a custom taxonomy. The result of the survey does
not actually change anything in the signup flow at this time, but can be used to
make other decisions later and may eventually influence the themes offered
during signup.

This tests three variants:

For all the users directed to the `vert-blog` or `vert-site` flow (this part is
taken care of by the homepage, which is not part of Calypso):

- 12% will see no survey and will see the normal signup flow instead.
- 44% will see the Verticals survey as the first step in the signup flow, with only one tier of categories.
- 44% will see the Verticals survey as the first step in the signup flow, with two tiers of categories.

In terms of the total number of users, this will amount to 10%/40%/40%
respectively because 10% of users will not be sent to these flows at all.