Signup: Verticals Survey: Add tracks event for final survey choice
Adds the event `calypso_survey_category_chosen` because it's difficult to
determine what category users finally choose otherwise. This is because if the
user is in the two tier test, they can click several first tier categories
before choosing one; if the user is in the one tier test, they can click only
once to choose a category. This event will represent the user's final choice.