Reader: Remove the exit routes that are killing j/k in full post view.
This is nasty. Because of how we render the full post view with a Dialog, the instance has a really odd lifetime. Since it's an overlay, when we leave the route, we should tear it down, but we also support going from one full post view to another. This means we would have a weird flash of transition as we moved if we tore down and spun up a new full post view on that move. Instead, we try to catch all the places we might go from the full post view and add a handler there to tear down the full post view.

This works great, except for Edit Post from the full post view. Find a post you can edit, go into full post view, then pick Edit Post from the options. The editor will spin up underneath the full post view and the url will change.

Ideally, exit routes would be passed the current route _and the next route_ so that we could be smart and only tear down the full post view when we're really leaving the URL space it covers.