Signup: Verticals Survey: Fix redirect-after-survey bug
Fixes #1310

The redirect was happening because in some browsers (iOS Chrome, Mac OS Firefox)
the `#` href in the `Card` components was redirecting the user back to the
survey step after the click handler had moved on to the next signup step. The
`#` href is necessary to force the `Card` components to display as clickable.

This fixes the issue by calling `preventDefault` and `stopPropagation` in the
click handler.