Importer: Poll from `v1.1` API enpoint
**Warning** This is a breaking build!

There are two pieces of this PR that break things:
 - The `v1.1` endpoint only exists in a patch - D398, but the `v1`
   endpoint _does_ exist but returns incompatable data. Since the
endpoints decide to send the wrong thing anyway, this will break unless
`public-api.wordpress.com` is sandboxed to a machine with the patch.
 - The polling automaticallly overwrites any local state, meaning that
   if you try to move forward with the import process your changes will
likely get wiped out before you are ready to go.

_*So what the heck is the PR for, anyhow?*_

This is here to start adding in the polling functionality without
requiring a big friendly merge. After this PR gets baked in, we should
be able to progress by intelligently merging local changes to the
importer states with those coming from the API.

**Testing**

Navigate to **My Sites** > **Settings** >> **Import** and start an
import by uploading a file. If you have previously done this, there is
likely an import waiting on the server and it will appear after about
five seconds. If not, then after your upload finishes (maybe before?)
the import from the server should appear. You should be able to navigate
away and come back and find things the same. You should be able to close
your browser session and come back and pick up where you left off (after
the delay in getting data from the API).