Interval: Run the passed action as soon as the component loads
Previously the interval ran the given action only on the timer, but this
meant that there would be an unexpected delay when loading a component
until the action ran, especially if it were on the EVERY_MINUTE period.

Now, for a single exception, when the `<Interval />` component loads, it
will run the given action immediately _and_ schedule it in the run
queue.