Show selected site first in site selector
Especially when adding a new post I almost always want to post to the
current site I have worked with – looked at posts, stats, etc.

It doesn’t make much of a difference in the sidebar picker, so I didn’t
exempt it from the change.

Note: in some cases (opening the reader directly) we don’t have the
selected site yet and we can’t show it. For now I decided not to fetch
it if missing.