Interval: Immediately execute `onTick()` in `start()`
Previously `<Interval />` was modified to run `onTick()` on component
mount, but sometimes a user would expect it to run immediately in other
circustances, such as when changing the interval period from one value
to another.

Now, the component will run `onTick()` whenever it is changed or
whenever the interval period is changed.