Gridicons: Adjust offset-fix to target the subgroup.
This PR wraps all the Gridicon SVGs in `<g>` subgroups, and moves the "offset-fix" to that instead. This has the primary benefit of compartmentalizing the fix itself further, since it's now on a child element of the SVG, hence allowing transforms on the SVG itself.

As a side-effect, it also means the offset-fix will work in situations where it previously didn't, due to how Chrome treats inline blocks.

The offset-fix primarily applies to gridicons set in the non-standard 18px size. Here's how an 18px icon looks *without* the offsetfix:

https://cloudup.com/csQDdo_Bh8f

Here's how that same icon looks *with* the offset fix applied.

https://cloudup.com/c4bSslRn2Sv