Make node-version an order-only dependency
This allows the dependency to run, but not foul the dep tree. With it as a normal dep, it would cause any target that depended upon it to re-run, every time it was invoked. This made a target like build-css rerun for every http request, which slowed things down considerably.

See https://www.gnu.org/software/make/manual/html_node/Prerequisite-Types.html for more info on order-only deps