Importer: Ignore stopped imports from the API
Previously when calling the API for getting the list of current imports,
if one had been previously cancelled, it would still show up in the
results from the API call. This led to issues in the UI because it
thought that there were still imports in progress.

This patch changes the behavior so that previously cancelled or stopped
import sessions get ignored in the UI.