enabling devdocs in all environments
There isn't much point in reserving this section to development and wpcalypso. wpcalypso is public anyway, so people can get to it... it's not like there's anything to hide. And enabling this in production isn't going to all of a sudden make it appear for regular users. It just gives us a less awkward place to point people to see some of our documentation and to play with our available components.