Add a marker to the geocoded location.
Based on the Google Static Maps Developer Guide there is an optional 'markers' feature parameter to define a marker to be attached to the map image at the specified location.

For a single marker the 'center' parameter can be left off and the map will be implicitly centered around the marker.

Note: having a marker defined, 'zoom' is not mandatory either but based on my quick testing the implicit zoom level is many times too high so I have not changed the original zoom level 8 in the code.