Importer: Fix an issue starting a new import (missing site ID)
As soon as the API integration began, small consistency issues between
the UI and the API started creeping in.  In this case, it was impossible
to start a new import session because that session depends on both the
type of the importer (already handled) and the site id for which it
belongs.

Previously that site id was missing but this PR adds it back in.