Editor: Contact Form Dialog Redux Store ༼ つ ◕_◕ ༽つ
Redux implementation to manage the state of the dialog window inside the TinyMCE plugin. This extends the editor reducer adding basic actions to load state from the editor content and to manage form fields.
The state is loaded/preserved from the editor content by serializing/deserializing the corresponding [contact-form] shortcode from the editor content.

TinyMCE command names changed to bring them up to standard with the rest of the plugins.