Update Gridicons, add nuance to offset-fix
The offset-fix nudges vectors inside the gridicon icons by subpixels in order to enhance crispness when used at 18 or 36px sizes.

This PR further enhances that by ambiguating whether an icon needs nudging only horizontally, vertically, or both. It also includes an updated list of which icons need either.

This has been tested in Calypso and works as intended.