Framework: remove unnecessary global constant for webpack-dev-server config
The path to the socket.io endpoint does not need to know the host information so the relative path `/` can be used in place of `http://calypso.localhost:3000` with the same effect.
The advantage of this change is the ability to run calypso in development mode using any hostname and port we want.