Themes: calc num columns for a smooth scroll
The number of theme screenshots per row in the Showcase is dynamic
and depends on the viewport size, yet the `<InfiniteList />` `itemsPerRow`
prop is currently hardcoded to `2`. `<InfiniteList />` uses this
prop to calculate spacer heights, so when it does not match with
the actual number of items per row, the scroll is glitchy.

This change adds a pragmatic calculation for that prop.