Importer: Fix bug where importer wouldn't enable
Previously, when visiting the import page for a site that had never had
an attempted import already, the importer would never enable itself,
remaining in the disabled state. The problem stemmed from the fact that
the API was sending a blank body, which caused a `TypeError` in the
`fromApi` call, halting the process which would otherwise have hydrated
the store, that is, indicate that a successful API call had been made.

Now the fetching logic maps over the response as a list so that null
responses like this won't break the importer.

This also includes a minor change to the store, adding a reset option to
support testing, and also some tests to validate these behaviors.