ThemesSiteSelectorModal: Get rid of broken message
We were displaying a "You will be able to buy the design after the preview" message
when invoking the customizer for premium theme preview. We no longer use the customizer
for preview, but the web preview for the generic demo site and thus don't invoke
the modal for preview anymore, so this message isn't shown any longer and can
hence be removed.

@folletto: Or do we want to keep it in the site selector modal that appears if
you click on 'Try & Customize' inside the theme preview?