Fix bug in importer - blank list after upload finishes
Previously the importer would flash a blank list right after the
uploaded file finished. This would disappear after a moment when the
polling system got an update back form the API.

The cause for this bug was the fact that the site id wasn't being set
for the response - or that this site id was missing from the API
response. Therefore, when it came back from the upload request, the
importer was effectively removed since it wouldn't have been available
in any site slot.

Now, the site id is being added back into the response so that it files
appropriately in the store. To accomplish this, a few changes and/or
improvements have been made:

 - The uploading function in `wpcom-undocumented` now returns a promise
   so that the code can be a little clearer in `actions.js`
 - The response is transformed in the upload file instead of the store
   for `finishUpload`