Reader: Support an `at` param to set where a stream starts.
This patch teaches all of the reader streams to start at a certain date. This helps when debugging streams that have a lot of volume, like tag streams, by allowing you to start at a particular date.

The at param will take any ISO-8601 compliant date, with timezone, and translate it to a UTC-based date before passing it down to the API.