Reader: Add an icon to the Visit and Share links on cards.
The Visit icon indicates that the click will take the user offsite.
The Share icon is because it looked lonely without it.

Also moved the click handler on the Share button up to the container, which matches where we apply the highlight and the pointer. No more just missing the click and moving to full post instead of the Sharing menu.

Last, stopped using onTouchTap, in favor of onClick, because onTouchTap doesn't seem capable of actually canceling an event properly, as least not in a manner that the card's click handler can understand and ignore.

Fixes #2623