Invites: Refactors match email; UI and style updates
Among other visual updates, for logged out users, do not show an action. For logged in users, "Switch Accounts" should log a user out and redirect to logged out flow for invites to an unknown user. For logged in invites to a known user, clicking "Switch Accounts" takes user to WP.com login page with a redirect back to logged in invite flow.