Fix up how feeds get handled, teach Site some new tricks.
Feeds don't have a valid Site object, so introduce some hijinks to get things to work.

Also noticed that clicking on a Site in full post view was loading the actual site after a short delay. After a bunch of testing, turns out that using a onTouchTap handler by itself is not enough to cancel a click on an anchor. The actual click event needs to be prevented as well, not just the mouseup that onTouchTap ends up piggybacking on.

Site can now take an onClick in addition to the onSelect. The major purpose is to allow hosts to decide whether or not to cancel the click on the anchor (if one is present).