Plugin: Improve single plugin view loading.
I updated the plugin exists method to return 'unknow'. In that cases
where do don't have enough information to determine if a plugin exists
or not.
Such as if the plugin is not on .org or if it is not part of any plugin
site.
This improved the loading of the single plugin page.