Reader: Switch to using page-notifier to decide when to remove the FPV.
This moves page-notifier out of lib/desktop and into lib/route. I checked to make sure the Desktop doesn't use it directly; the move appears to be safe.

It also teaches the reader controller to close the full post dialog when the route changes away from the FPV.

To test, open up a stream, hit j to pick the first post, then hit enter to open the post. Hit j again and the next post in the stream should load. Hit k and you should move back to the original post. Esc should close the FPV.

Also test using the back button in the browser and the back button at the top of the full post dialog.