Themes: Split main into single-site, multi-site,and logged-out
`my-sites/themes/main.jsx` has been overflowing with imports and methods, most
of which were used in only one or two out of the single-site, multi-site, and
logged-out scenario. Likewise, `theme-options` had to cater for all of those
three cases.

This commit hence breaks `main` into 3 components corresponding to each of those
cases and moves `getButtonOptions()` to each of them, removing conditionals that
would only serve the other cases. Because of fewer imports, methods, and
branches in each of those files, they should now be more legible.

To avoid redundancies, this PR also introduces `<ThemePreview>` for use with
each of the three components.