Plans: Prevent persistent loading state on `Plans` when `sites` is empty
Previously, if `sites-list` was empty on an initial load, it would not
trigger a render() on `Plans` when the sites eventually loaded. This is
because `Plans` used its `selectedSite` prop to determine which site ID
to request plans for, which is problematic:

- `selectedSite` is false when no site is selected or sites aren't
loaded.

- When sites do load, `getSelectedSite()` isn't called in the
controller, so `selectedSite` remains false until the given controller
method is called again.

This commit:

- Removes `selectedSite` as a prop of `Plans`, in favor of using
`sites.getSelectedSite()` inside that component.

- Updates `props.fetchSitePlans` to properly return the initial state
of a site's plan before they are loaded and only request site plans when
a selected site is present.