Importer: Replace progress bar witih spinner
Although the long-term plan is to provide feedback for how many objects
have already been imporer during an import session, that information is
currently unavailable. This means that the progress bar during the
import session remains blank.

Now, in the absence of that progress information, a spinner shows up in
its place to indicate that the process is currently working but not yet
finished.

It would have been possible to strip out of the progress bar, but I have
chosen to leave it in because it is already working given the
appropriate data from the server, and because we plan on adding that
data. Had this not already been tested with mock data, I would have
removed the code. By leaving it in, we immediately benefit when that
data becomes available and also have the added benefit that for certain
importers where a progress indication is unavailable (stream importers)
this should automatically show the right component for the task.

Fixed a bug where the spinner would show if the upload was complete.

Also rearranged some of the logic to meet newer Calypso standards

Importer: Fix two typos

Previously the importer would jump back to the UPLOADING state after
actually starting an import. This was a typo wherein the store was
setting the wrong state. Now it is fixed.

Also fixed a small piece that could have been an issue using `>=`
instead of simply checking for the presence of a progress value in the
import process. The bug could have appeared in briefly flashing the
spinner even if a progress update were being sent. This has now been
precluded.