Importer: Report progress
This takes possible progress values from the server and renders them in
the progress bar. It additionally writes a message below the progress
bar to indicate how many resources are still importing to give a better
indication to the user that things are still moving.

Move progress functionality outside of React component

Previously this information was contained in the React object, but
there's no need to do this as it doesn't rely on the instance `this` or
props or state.

The functions returning information about progress are now fat-arrow
functions defined above the React component.

Additionally, a few fixes have been made to the way they operate:

 - Previously the progress appeared to be present even if the
   `completed` property was missing. This resulted in a message saying
   "Waiting on NaN resources to import." The new function handles the
   multiple ways the API might currently report progress.

Formatting

Move import message getter into a function

Lock imports after clicking the done or cancel button

Previously, updates from the server/API could have overwritten changes
in the client after finishing or cancelling an import session.

Now changes for a given session are locked on completion to prevent
further updates from interfering.

Hook in error handling from API

Previously the API was returning a failure state for imports when the UI
wasn't responding or displaying that.

Now, the UI shows the unspecified error message (because no explanation
was coming from the API) and gives the user the option to cancel the
import.

Remove period from status message