Importer: Remove testing apparatus to flip through mock states
During the development of the importer we had a dropdown listing
possible states for the UI. As the API starts to provide real data,
testing can be accomplished with real data.

Additionally, the mock states started to conflict with the real data
anyway, so it is time this goes.

Remove Makefile because no tests exist