Improve check for progress information
Previously some import progress information was coming from the server
as `{ total: 0, completed: 0 }`. This didn't indicate any progress and
was being properly filtered out with the following check:

```js
if ( ! progress.completed )
```

Unfortunately that meant that the check also prevented showing a
progress bar when the count was zero.

Now, the check compares both the total and the completed properties to
make sure that both the completed property exists and that the totals
are not zero.