Themes: Use reduxStore in logged-out-design layout
Props @ockham

This allows the `is-section-...` class to be set properly
at the client end on a route change. The `ui` section
of the redux state is bootstrapped to the client for
the initial render.

(Only bootstrapping that section for now because
other sections of the tree use immutable maps,
which require special handling to hydrate.)