Importer: Allow manually expiring an import session.
Depends on D984

After an import finishes, the information representing its session
continues to live in the server and is passed back to the UI with the
successful state. This caused problems because the import screen would
show stale data for imports that were expired.

Now, by clicking the "Done" button we send a request to the server to
expire the import session so that it will get the new
"IMPORT_STATUS_EXPIRED" state and the UI can safely ignore that.

**To test**, apply D984 and sandbox `public-api.wordpress.com`, run an
import, and after it is finished, click on the "Done" button. In the
**master** branch, if you reload the page then the already finished
importer will reappear. In this branch, it shouldn't come back. Note
that you will need to test this across two separate import sessions to
confirm, or test in **master** first to see that the "Done" button is
ineffective.

N.B. Testing imports will take at least fifteen minutes to run due to
the way the current system handles imports. Do not be surprised when it
takes a long time.