Signup: Allow currentFlowName to be a function reference
If `currentFlowName` is a reference to a function instead of a string, it will
be called to determine the name of the current flow when there is no valid flow
passed to `utils.getFlowName()`.

This allows function calls with side-effects (e.g., using the `abtest()`
function) to be avoided if a specific flow is being requested.

Pass requested flowName to currentFlowName if it is a function