Plugins: added a tooltip to the 'add a plugin' icon-only button
Previously we were relying on only an icon to explain what that button did and it had no aria label for screen readers. It now has a tooltip when you hover and a screen reader label.

This issue was raised in #2257

Thanks to @ebinnion for helping me figure out how to make this work!