Add ping checks for desktop
Electron will periodically send pings to Calypso to make sure the server is still running. If we fail to get a response, Electron can detect that and handle accordingly.

In the same vein, Electron may die (e.g. `kill -9`) but Calypso can continue running, which is not great.

When a ping comes in (where the value of the ping is the frequency in ms for the pings), we ping back with that same frequency. We also start a killswitch that will exit the process if we don't get a ping back in a certain amount of time. This way, Calypso can kill itself if its parent process disappears.