move undefined site fix to getRecentlySelected()
This is an improvement to the fix in https://github.com/Automattic/wp-calypso/pull/3073. The issue is that your "recently selected" sites could potentially contain a site that has been removed from your list. And this had resulted in `getRecentlySelected()` including `undefined` for one of the sites in the array. Rather than protecting against having an undefined site in the component, it would be better if `getRecentlySelected` just filtered out invalid results.