Help: switch to using SectionHeader
I made a few changes, but all of them with the overall goal of using SectionHeader instead of a more custom gray-on-gray header in help search results. This acheives a couple of things: SectionHeader is the standard, I gave each section a bottom margin so it's easier to tell the various types of documentation apart, and allowed the footer link to look like a link instead of a text input. Also was able to remove a bunch of margin overrides so this is all a bit more sustainable.