Framework: Upgrade page to 1.6.4.
Fixes the busted wildcard params that we use in the reader to spin up the subscriptions store.

To see the diff, set a breakpoint on loadSubscriptions in reader/controller.js and load http://calypso.localhost:3000/read/blog/feed/10838258. With the fix, the break will trip and subs will be loaded. Without, they will not.

Another way to verify is to load that URL and check the network tab. You should see requests to /read/following/mine flowing out, but you will not in prod.