use polling to enable fetching when window (re)gains focus
I tried this in https://github.com/Automattic/wp-calypso/pull/2965 but I had to back it out before it hit production. It worked fine on mounting and for pagination, but I had removed the mechanism by which the component triggered a new query with different props. Woops.

So now I'm using a leading polling query, but I've added some protection against double-fetching on the fetchNextPage method.