handle deleted posts from a post-list request
In certain cases we get data for post-lists that overlaps with data already in memory. Because this data is fresher than the in-memory data, we may have learned that a post currently in the list should no longer be in the list. This commit handles this case and removes the post from the list.