Token Field: Extend component such that it can also receive an array of objects
For our work on people management, we have been working on adding statuses to the tokens in TokenField. Our initial pass at this was to use a callback. But, we ran into issues with the tokens updating properly after validation came back. While we got this working, we don't believe it to be an elegant solution.

So, this pass will allow us to create an array of objects that looks like `{ value: 'token value here', status: 'is-error' }` and send those in through the `value` prop.