Themes: Use real data for theme sheets
Wrap theme sheets with the theme-details component to use real data,
and fetch and cache data server-side for initial render.

Server-side data caching:

- if the theme's details aren't in the themeDetails cache, don't SSR &
  fire and forget a theme details XHR, which will renderToString when it
  has data for the theme.
- if the theme is in the cache, then we'll have markup for it, so render
  that

doing it this way means we don't have to block the request, but we lose
SSR for cache misses. i'm unsure of whether blocking on the response
will be a problem (need timing stats), if it is, memcache or a CDN would
be preferable.