Load the Customizer and post previews on sites that disallow iframes
We have some higher-security sites that always force the `x-frame-options: SAMEORIGIN`
HTTP header, breaking the Customizer and post previewing. We allow unsetting this header
through a nonce, which is now included in REST API responses for sites.

This may pave the way for increasing the security for everyone by enabling said HTTP
header on all WP.com sites.