People: Normalize invite create validations in the reducer
Previously, successful validations were normalized when updating the state in the invite form component. But, that could be an issue if the data is needed in more than one place.

This PR normalizes the success validations, which can sometimes be return as an object or an array, in the reducer so that components can simply retrieve the data.